pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "echo-agent"
version = "0.1.0"
description = "A modular AI agent framework with multi-channel support"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Echo Agent contributors" }]
keywords = ["ai", "agent", "assistant", "chatbot", "runtime"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "aiohttp>=3.9",
    "aiosqlite>=0.20",
    "croniter>=1.4",
    "numpy>=1.24",
]

[project.optional-dependencies]
openai = ["openai>=1.30"]
anthropic = ["anthropic>=0.40"]
bedrock = ["anthropic>=0.40", "boto3>=1.34"]
gemini = ["google-generativeai>=0.8"]
allproviders = ["openai>=1.30", "anthropic>=0.40", "boto3>=1.34", "google-generativeai>=0.8"]
vector = ["faiss-cpu>=1.7"]
container = ["docker>=7.0"]
vision = ["pillow>=10.0"]
process = ["psutil>=5.9"]
weixin = ["cryptography>=41.0"]
otel = ["opentelemetry-api>=1.20", "opentelemetry-sdk>=1.20", "opentelemetry-exporter-otlp>=1.20"]
tokenizers = ["tiktoken>=0.7"]
all = ["openai>=1.30", "anthropic>=0.40", "boto3>=1.34", "google-generativeai>=0.8", "pillow>=10.0", "psutil>=5.9", "faiss-cpu>=1.7", "cryptography>=41.0", "tiktoken>=0.7"]
dev = ["ruff>=0.4", "pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
echo-agent = "echo_agent.__main__:main"

[tool.hatch.build]
include = [
    "echo_agent/**/*.py",
    "echo_agent/config/**/*.yaml",
    "echo_agent/gateway/static/**/*",
    "skills/**/*.md",
    "skills/**/*.py",
    "skills/**/*.sh",
    "README.md",
    "scripts/install.sh",
]

[tool.hatch.build.targets.wheel]
packages = ["echo_agent"]

[tool.hatch.build.targets.wheel.sources]
"echo_agent" = "echo_agent"

[tool.hatch.build.targets.wheel.force-include]
"skills" = "echo_agent/_bundled/skills"

[tool.hatch.build.targets.sdist]
include = [
    "echo_agent/",
    "skills/",
    "scripts/install.sh",
    "README.md",
    "pyproject.toml",
]

[tool.ruff]
line-length = 120
target-version = "py311"
